cart icon updates, refactored cartcomponents
