+++ client/src/components/Cart.js
@@ -1,5 +1,7 @@
-//import icon from '../images/icons/cart-icon.png'
+
 import CartIcon from './CartIcon';
+import CartList from './cartlist';
+
 const Cart = ({ cartItems, setCart, cartOpen, setCartOpen}) => {
 
   if (cartOpen) {
@@ -14,47 +16,7 @@ const Cart = ({ cartItems, setCart, cartOpen, setCartOpen}) => {
   }
 
   return <CartIcon cartItems = {cartItems} setCartOpen = {setCartOpen}/>
-
   
-}
-
-const CartList = (props) => {
-
-  return (
-    <>
-    <ul className="Cart-box">
-      {props.cartItems.map((item) => {
-        return (
-
-          <li key={item.id}>
-            {item.name} - {item.qty} <button onClick={() => props.setCart(removeItem(item, props.cartItems))}>Remove</button>
-          </li>
-
-        );
-      })}
-    </ul>
-    <button id="buy-button" onClick={() => buy(props.cartItems)}>OSTA KAIKKI</button>
-    </>
-  );
-};
-
-function buy(items) {
-  /* OSTA */
-  console.log(items);
-}
-
-function removeItem(item, cartItems) {
-
-  /*HELP: removing works, HOWEVER: 
-  2. clicking a button wont update on click, only after updates after closing and opening the cart 
-  */
-
-
-  let removeThis = cartItems.indexOf(item)
-
-  if (removeThis > -1) {
-    return cartItems.splice(removeThis, 1)
-  }
 }
 
 export default Cart;
