+++ client/src/components/productDetails.js
@@ -1,9 +1,9 @@
-//import Cart from "./Cart";
 
-import { useState } from "react";
-import { useAuthContext } from "../hooks/useAuthContext";
 
-const ProductDetails = ({ product, visible, setCart, cartItems }) => {
+import CartComponent from "./cartComponent";
+
+const ProductDetails = ({ product, visible, setCart, cartItems,count,setCount }) => {
+
   if (!visible) return null;
 
   return (
@@ -39,69 +39,12 @@ const ProductDetails = ({ product, visible, setCart, cartItems }) => {
         item={product}
         setCart={() => setCart}
         cartItems={cartItems}
+        count = {count}
+        setCount = {setCount}
       />
     </div>
   );
 };
-const CartComponent = (props) => {
 
-  const [error,setError] = useState(null);
-  const { user } = useAuthContext();
-  
-  return (
-    <div className="add-to-cart-complex">
-      <input
-        type="number"
-        id="quantity"
-        className="qty-box"
-        min={1}
-        max={props.item.stock}
-        placeholder={1}
-      ></input>
-      <button
-        className="add-to-cart-button"
-        onClick={user ? () => props.setCart(addItem(props.item, props.cartItems)) : () => setError('You must be logged in')}
-      >
-        Add to cart
-      </button>
-      {error && <div className="error">{error}</div>}
-      <b id="total"></b>
-    </div>
-  );
-
-  function addItem(item, cartItems) {
-    let already = false;
-    let amount = parseInt(document.querySelector("#quantity").value);
-
-    console.log(cartItems);
-
-    if (!amount) amount = 1;
-    console.log("Määrä: " + amount);
-    console.log(cartItems);
-
-    cartItems.map((existing) => {
-      /* THIS SHIT DONT WORK */
-
-      console.log(existing);
-
-      if (item.name === existing.name) {
-        existing.qty = parseInt(existing.qty) + amount;
-        already = true;
-      }
-      return console.log("existing: " + existing.name);
-    })
-
-    /* THIS SHIT DO WORK */
-    if (!already) {
-      cartItems.push({
-        id: item.id,
-        name: item.name,
-        qty: amount
-      });
-    }
-
-    return cartItems;
-  }
-};
 
 export default ProductDetails;
